Trying to fix failing test
